update: updated typs for media fields

+++ services/media/errors.go
@@ -20,7 +20,7 @@ var (
 )
 
 type MediaTooLargeError struct {
-	Size int
+	Size uint64
 }
 
 func (e *MediaTooLargeError) Error() string {
